+++ src/new_client/mod.rs
@@ -1,8 +1,9 @@
 //! A proof-of-concept client based on smol.
 
+mod client;
+mod connection;
 mod decoder;
 mod encoder;
-mod connection;
 mod subscription;
 
 pub use connection::Connection;
