+++ src/new_client/subscription.rs
@@ -5,7 +5,7 @@ use futures::prelude::*;
 use smol::block_on;
 
 use crate::Message;
-use crate::new_client::connection::UserOp;
+use crate::new_client::client::UserOp;
 
 /// A subscription to a subject.
 pub struct Subscription {
@@ -16,7 +16,7 @@ pub struct Subscription {
     messages: mpsc::UnboundedReceiver<Message>,
 
     /// Enqueues user operations.
-    sub_ops: mpsc::UnboundedSender<UserOp>,
+    user_ops: mpsc::UnboundedSender<UserOp>,
 }
 
 impl Subscription {
@@ -24,12 +24,12 @@ impl Subscription {
     pub(crate) fn new(
         subject: &str,
         sid: usize,
-        sub_ops: mpsc::UnboundedSender<UserOp>,
+        user_ops: mpsc::UnboundedSender<UserOp>,
     ) -> Subscription {
         let (msg_sender, msg_receiver) = mpsc::unbounded();
 
         // Enqueue a SUB operation.
-        let _ = sub_ops.unbounded_send(UserOp::Sub {
+        let _ = user_ops.unbounded_send(UserOp::Sub {
             subject: subject.to_string(),
             queue_group: None,
             sid,
@@ -39,7 +39,7 @@ impl Subscription {
         Subscription {
             sid,
             messages: msg_receiver,
-            sub_ops,
+            user_ops,
         }
     }
 
@@ -53,7 +53,7 @@ impl Subscription {
 impl Drop for Subscription {
     fn drop(&mut self) {
         // Enqueue an UNSUB operation.
-        let _ = self.sub_ops.unbounded_send(UserOp::Unsub {
+        let _ = self.user_ops.unbounded_send(UserOp::Unsub {
             sid: self.sid,
             max_msgs: None,
         });
